+++ client/src/components/App.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import RelatedProducts from './RelatedProducts.jsx';
-import ReviewList from './ReviewList.jsx';
+import ReviewList from './RatingsReviews/ReviewList.jsx';
 import YourOutfit from './YourOutfit.jsx';
 import Overview from './Overview.jsx';
